backend/requirements.txt
bcrypt
Flask
Flask-cors
psycopg2
pyjwt
python-dotenv
requests
uwsgi

